# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RESMA Programmatic and Agency Report API",
    "version": "0.2.0",
    "summary": "Stable aggregate-only machine surface for analysis, agency report delivery, and AI workflows.",
    "description": "RESMA exposes aggregate-only observatory intelligence for AI clients, agency operators, and premium programmatic consumers. The API is a delivery channel for package-gated aggregate intelligence, not a raw contributor-data surface. Contributor-level feeds, per-user traces, and raw audience exports are intentionally excluded."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Programmatic API key. `Authorization: Bearer <api-key>` is also accepted."
      }
    },
    "parameters": {
      "format": {
        "name": "format",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "json",
            "llm",
            "markdown",
            "client-report"
          ],
          "default": "json"
        },
        "description": "Preferred response/export shape. Availability is package-dependent."
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "string"
          },
          "details": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": true
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          },
          "format": {
            "type": "string",
            "enum": [
              "llm",
              "markdown",
              "client-report"
            ]
          },
          "llm": {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "bullets": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "markdown": {
                "type": "string"
              },
              "followUpQuestions": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "caveats": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "additionalProperties": false
          },
          "export": {
            "type": "object",
            "properties": {
              "format": {
                "type": "string"
              },
              "content": {
                "oneOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "object",
                    "additionalProperties": true
                  }
                ]
              }
            },
            "additionalProperties": true
          }
        },
        "required": [
          "success",
          "data"
        ],
        "additionalProperties": true
      }
    }
  },
  "paths": {
...
    "/reports/packages/me": {
      "get": {
        "tags": [
          "Agency Reports"
        ],
        "summary": "Get the authenticated user package entitlements",
        "operationId": "getReportPackageAccess",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Package access metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/reports/presets": {
      "get": {
        "tags": [
          "Agency Reports"
        ],
        "summary": "List saved agency report presets",
        "operationId": "listAgencyReportPresets",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Saved presets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Agency Reports"
        ],
        "summary": "Create an agency report preset",
        "operationId": "createAgencyReportPreset",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 120
                  },
                  "reportType": {
                    "type": "string",
                    "enum": [
                      "AUDIENCE_OPPORTUNITY_BRIEF",
                      "COMPETITOR_REACH_SNAPSHOT",
                      "RECOMMENDATION_GAP_REPORT"
                    ]
                  },
                  "reportConfig": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "allowedExportFormats": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "json",
                        "llm",
                        "markdown",
                        "client-report"
                      ]
                    }
                  }
                },
                "required": [
                  "name",
                  "reportType",
                  "reportConfig"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Preset created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/reports/presets/{presetId}/run": {
      "post": {
        "tags": [
          "Agency Reports"
        ],
        "summary": "Generate a reproducible report run from a preset",
        "operationId": "runAgencyReportPreset",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "presetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Report run created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/reports/runs": {
      "get": {
        "tags": [
          "Agency Reports"
        ],
        "summary": "List generated report runs",
        "operationId": "listAgencyReportRuns",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Report runs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/reports/runs/{reportRunId}": {
      "get": {
        "tags": [
          "Agency Reports"
        ],
        "summary": "Get a stored report run with operator payload",
        "operationId": "getAgencyReportRun",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "reportRunId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/reports/runs/{reportRunId}/export": {
      "get": {
        "tags": [
          "Agency Reports"
        ],
        "summary": "Export a stored report run in a package-allowed format",
        "operationId": "exportAgencyReportRun",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "reportRunId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/format"
          }
        ],
        "responses": {
          "200": {
            "description": "Curated aggregate report export",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/reports/runs/{reportRunId}/shares": {
      "get": {
        "tags": [
          "Agency Reports"
        ],
        "summary": "List active share links for a report run",
        "operationId": "listAgencyReportShares",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "reportRunId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active share links",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Agency Reports"
        ],
        "summary": "Create a read-only shared report link",
        "operationId": "createAgencyReportShare",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "reportRunId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "maxLength": 160
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Share link created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/reports/shares/{shareId}": {
      "delete": {
        "tags": [
          "Agency Reports"
        ],
        "summary": "Revoke a share link",
        "operationId": "revokeAgencyReportShare",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "shareId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Share revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/shared-reports/{shareToken}": {
      "get": {
        "tags": [
          "Shared Reports"
        ],
        "summary": "Read a shared aggregate report export",
        "operationId": "getSharedAgencyReport",
        "parameters": [
          {
            "name": "shareToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/format"
          }
        ],
        "responses": {
          "200": {
            "description": "Shared read-only report export",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Shared report not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/runs/{reportRunId}/export": {
      "get": {
        "tags": [
          "Programmatic Analysis"
        ],
        "summary": "Read a saved report export through an API key",
        "operationId": "getProgrammaticAgencyReportExport",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "reportRunId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/format"
          }
        ],
        "responses": {
          "200": {
            "description": "Programmatic aggregate report export",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Package or scope does not allow report export",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}